beta

+++ db/migrate/20181104140035_create_ilearn_receits_revenues.rb
@@ -3,10 +3,10 @@ class CreateIlearnReceitsRevenues < ActiveRecord::Migration[5.2]
     create_table :ilearn_receits_revenues do |t|
       t.bigint :district_number
       t.string :district_name
-      t.integer :state_amount
-      t.integer :local_amount
-      t.integer :federal_amount
-      t.integer :total_amount
+      t.bigint :state_amount
+      t.bigint :local_amount
+      t.bigint :federal_amount
+      t.bigint :total_amount
       t.string :data_source_url
       t.string :scrape_dev
