+++ db/migrate/20181104140102_create_ilearn_expenditures_disbursements.rb
@@ -3,11 +3,11 @@ class CreateIlearnExpendituresDisbursements < ActiveRecord::Migration[5.2]
     create_table :ilearn_expenditures_disbursements do |t|
       t.bigint :district_number
       t.string :district_name
-      t.integer :instruction_amount
-      t.integer :general_administration_amount
-      t.integer :suppert_services_amount
-      t.integer :other_amount
-      t.integer :total_amount
+      t.bigint :instruction_amount
+      t.bigint :general_administration_amount
+      t.bigint :suppert_services_amount
+      t.bigint :other_amount
+      t.bigint :total_amount
       t.string :data_source_url
       t.string :scrape_dev
