+++ db/migrate/20181104140143_create_ilearn_tax_information.rb
@@ -3,8 +3,8 @@ class CreateIlearnTaxInformation < ActiveRecord::Migration[5.2]
     create_table :ilearn_tax_information do |t|
       t.bigint :district_number
       t.string :district_name
-      t.integer :real_eav
-      t.integer :real_per_pupil
+      t.bigint :real_eav
+      t.bigint :real_per_pupil
       t.integer :statewide_eavpp_rank
       t.string :formula_type
       t.integer :total_tax_rate
